📦 chore: 역경과 고난
