+++ src/components/dashboard/CreateCardModal/index.tsx
@@ -27,7 +27,7 @@ export type CreateCardInputs = {
   assignee: number;
   title: string;
   description: string;
-  dueDate: Date;
+  dueDate?: Date;
   tags?: string[];
   image?: string;
 };
@@ -56,7 +56,7 @@ export default function CreateCardModal({
       title,
       description,
       tags,
-      dueDate: formatDate(dueDate),
+      dueDate: dueDate && formatDate(dueDate),
       imageUrl: image,
     };
